requirements.txt
python-dotenv
numpy
pandas
ta
backtesting
ccxt
scikit-learn
flask
keyring
pytest